restructured some tests
